# openapi/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Parks API
x-yc-apigateway:
  service_account_id: ${PARKS_API_SA_ID}

paths:
  /{file+}:
    options:
      parameters:
        - name: file
          in: path
          description: greedy param
          required: false
          schema:
            type: string
      description: Greedy options
      operationId: greedyOptions
      x-yc-apigateway-integration:
        type: dummy
        http_code: 200
        content:
          "*": ""
        http_headers:
          Allow: OPTIONS, GET, HEAD, POST
          Access-Control-Allow-Origin: '*'
          Access-Control-Allow-Headers: Content-Type, Authorization
  /parks:
    get:
      description: Возвращает парки в порядке убывания голосов
      operationId: getParks
      #security:
      #  - httpBearerAuth: [ ]
      parameters:
        - name: limit
          in: query
          description: maximum number of results to return
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: parks response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Park'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${PARKS_API_CONTAINER_ID}
    post:
      description: Upsert parks
      operationId: postParks
      #security:
      #  - httpBearerAuth: [ ]
      requestBody:
        description: Park to upsert
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Park'
      responses:
        '200':
          description: park response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Park'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${PARKS_API_CONTAINER_ID}
  /parks/{id}:
    parameters:
      - name: id
        in: path
        description: ID of park to delete
        required: true
        schema:
          type: integer
    get:
      description: Returns park based on the ID supplied
      operationId: getParkById
      #security:
      #  - httpBearerAuth: [ ]
      responses:
        '200':
          description: park response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Park'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${PARKS_API_CONTAINER_ID}
    delete:
      description: deletes a single park based on the ID supplied
      operationId: deleteParkById
      #security:
      #  - httpBearerAuth: [ ]
      responses:
        '204':
          description: park deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${PARKS_API_CONTAINER_ID}
  /users/{userId}/votes:
    post:
      description: Post vote
      operationId: postVote
      #security:
      #  - httpBearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          schema:
            type: string
      requestBody:
        description: Vote to post
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vote'
      responses:
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-yc-apigateway-integration:
        type: cloud_datastreams
        action: PutRecord
        stream_name: ${PARKS_DATABASE_PATH}/votes-stream
        partition_key: '{userId}'
    get:
      description: Get vote
      operationId: getVote
      #security:
      #  - httpBearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          schema:
            type: string
        - name: parksId
          in: query
          description: ID of park
          required: true
          schema:
            type: number
      responses:
        '200':
          description: vote response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${PARKS_API_CONTAINER_ID}
components:
  schemas:
    Park:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: integer
        title:
          type: string
        poster_path:
          type: string
        vote_count:
          type: integer
        vote_average:
          type: number
        overview:
          type: string

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Vote:
      type: object
      required:
        - park_id
        - user_id
        - value
      properties:
        park_id:
          type: integer
        user_id:
          type: string
        value:
          type: integer
  securitySchemes:
    httpBearerAuth:
      type: http
      scheme: bearer
#      x-yc-apigateway-authorizer:
#        type: function
#        function_id: ${AUTHORIZER_FUNCTION_ID}
#        authorizer_result_ttl_in_seconds: 7200